feat: add confirm dialog for deleting projects

+++ src/app/pages/project-page/project-page.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ProjectService } from '../../features/project/project.service';
 import { MatDialog } from '@angular/material';
 import { DialogCreateProjectComponent } from '../../features/project/dialogs/create-project/dialog-create-project.component';
+import { DialogConfirmComponent } from '../../ui/dialog-confirm/dialog-confirm.component';
 
 @Component({
   selector: 'project-page',
@@ -30,6 +31,18 @@ export class ProjectPageComponent implements OnInit {
   }
 
   remove(projectId) {
-    this.projectService.remove(projectId);
+    this._matDialog.open(DialogConfirmComponent, {
+      restoreFocus: true,
+      data: {
+        okTxt: 'Delete',
+        cancelTxt: 'Cancel',
+        message: `Are you sure you want to delete this project?`,
+      }
+    }).afterClosed()
+      .subscribe((isConfirm: boolean) => {
+        if (isConfirm) {
+          this.projectService.remove(projectId);
+        }
+      });
   }
 }
